+++ src/DSP/HCVRandom.h
@@ -1,6 +1,8 @@
 #pragma once
 
 #include "Gamma/rnd.h"
+#include "HCVFunctions.h"
+#include "math.hpp"
 
 class HCVRandom
 {
@@ -30,7 +32,7 @@ public:
     
     float nextGaussian() const
     {
-        return gam::rnd::gaussian(gamRand);
+        return gam::rnd::gaussian(gamRand) * 0.3;
     }
 
     bool nextProbability(float _prob) const
@@ -50,4 +52,22 @@ public:
 
 private:
     gam::RNGMulCon gamRand;
+};
+
+class HCVGrayNoise
+{
+public:
+	float operator()()
+	{
+		auto newNoise = int(randGen.nextFloat() * 31.0) & 31;
+		newNoise = 1 << newNoise;
+		lastNoise = newNoise ^ lastNoise;
+
+		const auto output = lastNoise * 4.65661287308e-10f;
+		return rack::math::clamp(uniToBi(output), -1.0f, 1.0f);
+	}
+
+private:
+	int lastNoise = 0;
+	HCVRandom randGen;
 };