+++ src/DSP/HCVShiftRegister.h
@@ -2,6 +2,8 @@
 
 #include "dsp/digital.hpp"
 #include <vector>
+#include "HCVRandom.h"
+#include "HCVFunctions.h"
 
 template <typename T = float>
 class HCVShiftRegister
@@ -77,4 +79,30 @@ public:
 
 private:
     float runglerOut = 0.0f;
+};
+
+class HCVLFSRNoise : public HCVShiftRegister<float>
+{
+public:
+    float operator()()
+    {
+        advanceRegister(randGen.nextBoolean());
+
+        float noiseValue = dataRegister[0] ? 1.0f : 0.0f;
+        noiseValue += dataRegister[1] ? 2.0f : 0.0f;
+        noiseValue += dataRegister[2] ? 4.0f : 0.0f;
+        noiseValue += dataRegister[3] ? 8.0f : 0.0f;
+        noiseValue += dataRegister[4] ? 16.0f : 0.0f;
+        noiseValue += dataRegister[5] ? 32.0f : 0.0f;
+        noiseValue += dataRegister[6] ? 64.0f : 0.0f;
+        noiseValue += dataRegister[7] ? 128.0f : 0.0f;
+
+        noiseValue = noiseValue/255.0f;
+        noiseValue = uniToBi(noiseValue);
+
+        return noiseValue;
+    }
+
+private:
+    HCVRandom randGen;
 };