+++ src/HetrickCV.hpp
@@ -40,6 +40,7 @@ extern Model *modelChaos1Op;
 extern Model *modelChaos2Op;
 extern Model *modelChaos3Op;
 extern Model *modelChaoticAttractors;
+extern Model *modelClockedNoise;
 extern Model *modelComparator;
 extern Model *modelContrast;
 extern Model *modelCrackle;